+++ .github/scripts/behavior_test/test_plan.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -17,7 +16,7 @@
 # under the License.
 
 import unittest
-from behavior_test_plan import plan
+from plan import plan
 
 
 class BehaviorTestPlan(unittest.TestCase):
